Extract check_and_fix function
